added details admonition
